requirements.txt
flask==2.3.3
tensorflow==2.16.2
pillow==10.0.0
astropy==5.3.0
numpy==1.24.3
